+++ packages/dht/src/dht/routing/Router.ts
@@ -42,9 +42,21 @@ interface ForwardingTableEntry {
     peerDescriptors: PeerDescriptor[]
 }
 
+interface IRouterFunc {
+    doRouteMessage(routedMessage: RouteMessageWrapper, mode: RoutingMode): RouteMessageAck
+    send(msg: Message, reachableThrough: PeerDescriptor[]): Promise<void>
+    checkDuplicate(messageId: string): boolean
+    addToDuplicateDetector(messageId: string, senderId: string, message?: Message): void
+    addRoutingSession(session: RoutingSession): void
+    removeRoutingSession(sessionId: string): void
+    stop(): void
+}
+
+export interface IRouter extends Omit<IRoutingService, 'findRecursively'>, IRouterFunc {}
+
 const logger = new Logger(module)
 
-export class Router implements Omit<IRoutingService, 'findRecursively'> {
+export class Router implements IRouter {
     private readonly config: RouterConfig
     private readonly forwardingTable: Map<string, ForwardingTableEntry> = new Map()
     private ongoingRoutingSessions: Map<string, RoutingSession> = new Map()
@@ -70,7 +82,7 @@ export class Router implements Omit<IRoutingService, 'findRecursively'> {
                 requestId: v4(),
                 destinationPeer: forwardingPeer,
                 sourcePeer: this.config.ownPeerDescriptor!,
-                reachableThrough: [],
+                reachableThrough,
                 routingPath: []
             }
             this.doRouteMessage(forwardedMessage, RoutingMode.FORWARD)
@@ -83,7 +95,7 @@ export class Router implements Omit<IRoutingService, 'findRecursively'> {
                 reachableThrough,
                 routingPath: []
             }
-            this.doRouteMessage(routedMessage, RoutingMode.FORWARD)
+            this.doRouteMessage(routedMessage, RoutingMode.ROUTE)
         }
     }